chore(cmd/run): resource type package reorg
